refactor: delegate app_state methods to feature modules
- Create update_suggestions_from_app() in autocomplete module
- Create insert_suggestion_from_app() in autocomplete/insertion module
- Create update_tooltip_from_app() in tooltip module
- Create update_stats_from_app() in stats module
- Update App methods to be one-line delegation calls
- Move feature-specific tests to their respective modules
- Move property-based tests to feature modules

All 774 tests pass (766 unit + 8 integration).
