+++ src/stats.rs
@@ -9,3 +9,4 @@ mod types;
 
 // Re-export public types
 pub use stats_state::StatsState;
+pub use stats_state::update_stats_from_app;
